// src/dorc-web/src/apis/dorc-api/models/PagedDataFilter.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface PagedDataFilter
 */
export interface PagedDataFilter {
    /**
     * @type {string}
     * @memberof PagedDataFilter
     */
    Path?: string | null;
    /**
     * @type {string}
     * @memberof PagedDataFilter
     */
    FilterValue?: string | null;
}