// src/dorc-web/src/apis/dorc-api/models/GetDaemonAuditListResponseDto.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type {
    DaemonAuditApiModel,
} from './';

/**
 * @export
 * @interface GetDaemonAuditListResponseDto
 */
export interface GetDaemonAuditListResponseDto {
    /**
     * @type {number}
     * @memberof GetDaemonAuditListResponseDto
     */
    CurrentPage?: number;
    /**
     * @type {number}
     * @memberof GetDaemonAuditListResponseDto
     */
    TotalItems?: number;
    /**
     * @type {number}
     * @memberof GetDaemonAuditListResponseDto
     */
    TotalPages?: number;
    /**
     * @type {Array<DaemonAuditApiModel>}
     * @memberof GetDaemonAuditListResponseDto
     */
    Items?: Array<DaemonAuditApiModel> | null;
}